sort courses by course name
there seems to be no way to get angular to sort the options defined by
ng-options if the data source is an object.  however, the reasons for
needing to have the data source be an object no longer apply.  let's
just use a list of courses as the data source instead, since they come
sorted already from the server.