seemingly unnecessary .all() speeds up large querysets
i have no idea why, but the large term i tested with (7k+ courses) timed
out without this.  it seems like the pagination logic doesn't affect the
actual db query if we just do

    model.objects.filter(...)

but it does if we do

    model.objects.filter(...).all()

i'm *guessing* this has something to do with the queryset cache busting
metioned at

    https://docs.djangoproject.com/en/1.8/ref/models/querysets/#all

but that's just a guess.