canvas_account_course_admin project has not had a successful run in Jenkins since late March. There appears to be a persistent failure in the "course_info_remove_people_tests" -  it's failing due to "StaleElementReferenceException: Message: Element not found in the cache - perhaps the page has changed since it was looked up"
Error is relatively persistent;  the test users at which it failed on is not at all consistent (i.e. sometimes it fails on test 4, other times it fails on test 7).
This had been reproducible on Jenkins and locally.  Adding the code fixes this issue.   I have not been able to reproduce the stale_element error with this change locally and when I run it on Jenkins anymore.